requirements.txt
disnake==2.1.0
python-dotenv==0.17.0